build(core): fix some type error
