+++ packages/collapse/src/collapse.ts
@@ -1,0 +1,7 @@
+import { Emitter } from 'mitt'
+import { Ref } from 'vue'
+
+export interface CollapseProvider {
+  activeNames: Ref
+  collapseMitt: Emitter
+}
