+++ packages/color-picker/src/index.vue
@@ -306,15 +306,13 @@ export default defineComponent( {
 })
 </script>
 
-<style lang="scss" scoped>
+<style>
 .el-color-picker:focus {
   outline: none;
 }
 .hue-slider {
   float: right;
 }
-</style>
-<style>
 .el-popper.el-color-picker__panel {
   border-color: #ebeef5;
 }
