+++ packages/pagination/src/sizes.vue
@@ -22,7 +22,7 @@ import { defineComponent, watch, computed, ref } from 'vue'
 import ElSelect from '@element-plus/select/src/select.vue'
 import ElOption from '@element-plus/select/src/option.vue'
 import { t } from '@element-plus/locale'
-import { isEqual } from 'lodash'
+import isEqual from 'lodash/isEqual'
 import { usePagination } from './usePagination'
 
 export default defineComponent({
